Move abs to dedicated class
